Moved write-target related code out of pkg/beholder
