+++ pkg/capabilities/write_target/monitor/schema.go
