+++ pkg/capabilities/write_target/report/mercury/common/types.go
