+++ pkg/capabilities/write_target/report/platform/types.go
