+++ pkg/capabilities/write_target/report/platform/types_test.go
